Split App.tsx and gui.rs into smaller, focused modules
App.tsx (4,357 → 2,424 lines) and gui.rs (2,783 lines → 10-file gui/ directory)
were too large to easily read and maintain. This extracts types, utilities,
components, and feature modules into separate files with no behavior changes.
