+++ src/utils/version.ts
@@ -1,0 +1,9 @@
+export function isNewerVersion(current: string, latest: string): boolean {
+  const c = current.split(".").map(Number);
+  const l = latest.split(".").map(Number);
+  for (let i = 0; i < 3; i++) {
+    if (l[i] > c[i]) return true;
+    if (l[i] < c[i]) return false;
+  }
+  return false;
+}
